+++ src/app/cart/components/cart/cart.component.ts
@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+
+import { Observable } from 'rxjs';
 
 import { CartService } from '../../services';
 
@@ -8,13 +10,14 @@ import { CartProduct } from '../../../products/models';
   selector: 'app-cart',
   templateUrl: './cart.component.html',
   styleUrls: ['./cart.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CartComponent implements OnInit {
-  cartItems: CartProduct[];
+  cartItems$: Observable<CartProduct[]>;
 
   constructor(private cartService: CartService) { }
 
   ngOnInit() {
-    this.cartItems = this.cartService.getCartItems();
+    this.cartItems$ = this.cartService.cartItems$;
   }
 }